Cleanup: remove partial copy of Linux private header genhd.h
Move the only structure required to the file that uses it.
